server/requirements/dev.txt
django-debug-toolbar
model-bakery